+++ src/cn/shawn/iem/ReverseInfluenceSampling.java
@@ -111,7 +111,7 @@ public class ReverseInfluenceSampling {
 		long startTime = System.currentTimeMillis();
 		ArrayList<String> s = ris.calculateSourceSet(dirWgtGph, k, r);
 		long endTime = System.currentTimeMillis();
-		long runTimeSec = (endTime - startTime)/1000;
+		double runTimeSec = (endTime - startTime)/1000.0;
 		System.out.println("k = " + k + ", r = " + r + ", runtime = " + runTimeSec + " secs.");
 		System.out.println(s);
 	}
